Make `else` case required, add `unreachable` kw
Switch expressioned added a small issue. In the
example below, what should happen is a little
unclear:

var x = switch (3) {
  case 1 -> "2";
  case 2 -> "3";
};

Should it error? Return `null`?
Previously, if a case was unhandled, it would
throw a runtime error. But now it's been changed
to require an else statement, to be more
explicit. The previous behavior can be mimicked
by using the new `unreachable` keyword.

var x = switch (3) {
  case 1 -> "2";
  case 2 -> "3";
  else -> unreachable;
};